requirements.txt
PyYAML>=6.0
Jinja2>=3.1
jsonschema>=4.21
pydantic>=2.7
requests>=2.31
